in the Testimonials section
in the first article
replace the text with a blockquote with text I am completely blown away. Thanks to Techium, we've just launched our 5th website! and cite author Yuri Y.
in the second article
replace the text with a blockquote with text Thank you so much for your help. Techium company is awesome! and cite author Dorrie S.
in the third article
replace the text with a blockquote with text I love your system. Definitely worth the investment. I'd be lost without Techium company. and cite author Sven H.